requirements.txt
Django==1.11.29
psycopg2-binary
djangorestframework
djangorestframework-simplejwt
python-decouple